+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### IMPROVEMENTS
 
 - Linz will automatically look for `label` and `name` fields when trying to determine the `title` of a record. When producing the list of fields to include in the select portion of the query, it will now include those, if the model has fields of the same names.
+- Linz will not attempt to render fields in the `columns` list, if its value is `falsy`.
 
 ## v0.7.0.0 (3 March 2017)